Refactoring importing of CSV and XML File, Changed PlayerInfo Classes
There were major issues with the CSV file import (in the sense that it was completely wrong if teams are actually a thing) and the XML file import was completely inoperable. Both now work as expected. Additionally made some changes to the classes so that everything isn't just based off `PlayerInfo` but rather the other way around.

Also made the link to Bandai's website actually function. (Thanks to SO for that one.)

Started working on tournament penalties and staff listing.

TODO: Finish staff and penalties entry.
TODO: Team of two and three registration.
TODO: Create documentation
TODO: Deployment